Changed folder structure for tests based on module - auth, events, attendees
