+++ backend/src/tests/eventTests/userEventInvite.test.ts
@@ -1,5 +1,5 @@
 import request from 'sync-request-curl';
-import { port, url } from '../config.json';
+import { port, url } from '../../config.json';
 
 const SERVER_URL = `${url}:${port}`;
 const TIMEOUT_MS = 5 * 1000;
