+++ backend/src/tests/eventTests/userUpdateEvent.test.ts
@@ -1,11 +1,13 @@
 import request from 'sync-request-curl';
-import { port, url } from '../config.json';
+import { port, url } from '../../config.json';
+import { UpdateEvents } from '../../interfaces';
 
 const SERVER_URL = `${url}:${port}`;
 const TIMEOUT_MS = 5 * 1000;
 
 let token: string;
 let eventId: string;
+let updatedFields: UpdateEvents;
 beforeEach(() => {
   request('DELETE', SERVER_URL + '/clear', { timeout: TIMEOUT_MS });
   requestAuthRegister('Mubashir', 'Hussain', 'Abcdefg123$', 'example@gmail.com');
@@ -16,6 +18,15 @@ beforeEach(() => {
   const res1 = requestNewEvent(token, 'New Event', 'New Description', 'House', '31/08/2025', 10, 14);
   const data1 = JSON.parse(res1.body.toString());
   eventId = data1.eventId;
+
+  updatedFields = {
+    title: 'Different Event',
+    description: 'Different Description',
+    location: 'Different House',
+    date: '31/08/2025',
+    startTime: 10,
+    endTime: 14
+  };
 });
 
 afterEach(() => {
@@ -24,14 +35,14 @@ afterEach(() => {
 
 describe('Error Cases', () => {
   test('Invalid UserId Token', () => {
-    const res = requestEventDetails('InvalidToken', eventId);
+    const res = requestEventUpdate('InvalidToken', eventId, updatedFields);
     const data = JSON.parse(res.body.toString());
     expect(data).toStrictEqual({ error: expect.any(String) });
     expect(res.statusCode).toStrictEqual(401);
   });
 
   test('Invalid EventID', () => {
-    const res = requestEventDetails(token, 'InvalidEventId');
+    const res = requestEventUpdate(token, 'InvalidEventId', updatedFields);
     const data = JSON.parse(res.body.toString());
     expect(data).toStrictEqual({ error: expect.any(String) });
     expect(res.statusCode).toStrictEqual(400);
@@ -39,14 +50,22 @@ describe('Error Cases', () => {
 });
 
 describe('Success', () => {
-  test('Success', () => {
+  test('Successful Return type', () => {
+    const res = requestEventUpdate(token, eventId, updatedFields);
+    const data = JSON.parse(res.body.toString());
+    expect(data).toStrictEqual({});
+    expect(res.statusCode).toStrictEqual(200);
+  });
+
+  test('Successfully Updated', () => {
+    requestEventUpdate(token, eventId, updatedFields);
     const res = requestEventDetails(token, eventId);
     const data = JSON.parse(res.body.toString());
     expect(data.event).toStrictEqual({
       id: eventId,
-      title: 'New Event',
-      description: 'New Description',
-      location: 'House',
+      title: 'Different Event',
+      description: 'Different Description',
+      location: 'Different House',
       date: '31/08/2025',
       startTime: 10,
       endTime: 14,
@@ -75,6 +94,14 @@ const requestNewEvent = (token: string, title: string, description: string, loca
   return (request('POST', SERVER_URL + '/events/new-event', {
     headers: { Authorization: `Bearer ${token}` },
     json: { title, description, location, date, startTime, endTime },
+    timeout: TIMEOUT_MS
+  }));
+};
+
+const requestEventUpdate = (token: string, eventId: string, updatedFields: UpdateEvents) => {
+  return (request('PUT', SERVER_URL + `/events/update-event/${eventId}`, {
+    headers: { Authorization: `Bearer ${token}` },
+    json: updatedFields,
     timeout: TIMEOUT_MS
   }));
 };
